fix: generalize fix for swallowed hover events
Replaces the old drag hover fix (which fixed kindof the same thing)
with one that works in a general fashion only on element.* level.

Related to https://github.com/bpmn-io/bpmn-js/issues/1366
