+++ lib/features/dragging/index.js
@@ -1,15 +1,13 @@
+import HoverFixModule from '../hover-fix';
 import SelectionModule from '../selection';
 
 import Dragging from './Dragging';
-import HoverFix from './HoverFix';
+
 
 export default {
-  __init__: [
-    'hoverFix'
-  ],
   __depends__: [
-    SelectionModule
+    HoverFixModule,
+    SelectionModule,
   ],
   dragging: [ 'type', Dragging ],
-  hoverFix: [ 'type', HoverFix ]
 };
