+++ lib/features/hover-fix/HoverFix.js
@@ -20,13 +20,15 @@ var HIGH_PRIORITY = 1500;
  * 1) have a hover state after a successful drag.move event
  * 2) have an out event when dragging leaves an element
  *
- * @param {EventBus} eventBus
- * @param {Dragging} dragging
  * @param {ElementRegistry} elementRegistry
+ * @param {EventBus} eventBus
+ * @param {Injector} injector
  */
-export default function HoverFix(eventBus, dragging, elementRegistry) {
+export default function HoverFix(elementRegistry, eventBus, injector) {
 
   var self = this;
+
+  var dragging = injector.get('dragging', false);
 
   /**
    * Make sure we are god damn hovering!
@@ -58,80 +60,73 @@ export default function HoverFix(eventBus, dragging, elementRegistry) {
     }
   }
 
-  /**
-   * We wait for a specific sequence of events before
-   * emitting a fake drag.hover event.
-   *
-   * Event Sequence:
-   *
-   * drag.start
-   * drag.move >> ensure we are hovering
-   */
-  eventBus.on('drag.start', function(event) {
 
-    eventBus.once('drag.move', HIGH_PRIORITY, function(event) {
+  if (dragging) {
 
-      ensureHover(event);
+    /**
+     * We wait for a specific sequence of events before
+     * emitting a fake drag.hover event.
+     *
+     * Event Sequence:
+     *
+     * drag.start
+     * drag.move >> ensure we are hovering
+     */
+    eventBus.on('drag.start', function(event) {
+
+      eventBus.once('drag.move', HIGH_PRIORITY, function(event) {
+
+        ensureHover(event);
+
+      });
 
     });
-
-  });
+  }
 
 
   /**
-   * We make sure that drag.out is always fired, even if the
+   * We make sure that element.out is always fired, even if the
    * browser swallows an element.out event.
    *
    * Event sequence:
    *
-   * drag.hover
+   * element.hover
    * (element.out >> sometimes swallowed)
-   * element.hover >> ensure we fired drag.out
+   * element.hover >> ensure we fired element.out
    */
-  eventBus.on('drag.init', function() {
+  (function() {
+    var hoverGfx;
+    var hover;
 
-    var hover, hoverGfx;
+    eventBus.on('element.hover', function(event) {
 
-    function setDragHover(event) {
-      hover = event.hover;
-      hoverGfx = event.hoverGfx;
-    }
-
-    function unsetHover() {
-      hover = null;
-      hoverGfx = null;
-    }
-
-    function ensureOut() {
-
-      if (!hover) {
-        return;
-      }
-
-      var element = hover,
-          gfx = hoverGfx;
-
-      hover = null;
-      hoverGfx = null;
-
-      // emit synthetic out event
-      dragging.out({
-        element: element,
-        gfx: gfx
-      });
-    }
-
-    eventBus.on('drag.hover', setDragHover);
-    eventBus.on('element.out', unsetHover);
-    eventBus.on('element.hover', HIGH_PRIORITY, ensureOut);
-
-    eventBus.once('drag.cleanup', function() {
-      eventBus.off('drag.hover', setDragHover);
-      eventBus.off('element.out', unsetHover);
-      eventBus.off('element.hover', ensureOut);
+      // (1) remember current hover element
+      hoverGfx = event.gfx;
+      hover = event.element;
     });
 
-  });
+    eventBus.on('element.hover', HIGH_PRIORITY, function(event) {
+
+      // (3) am I on an element still?
+      if (hover) {
+
+        // (4) that is a problem, gotta "simulate the out"
+        eventBus.fire('element.out', {
+          element: hover,
+          gfx: hoverGfx
+        });
+      }
+
+    });
+
+    eventBus.on('element.out', function() {
+
+      // (2) unset hover state if we correctly outed us *GG*
+      hoverGfx = null;
+      hover = null;
+    });
+
+  })();
 
   this._findTargetGfx = function(event) {
     var position,
@@ -152,9 +147,9 @@ export default function HoverFix(eventBus, dragging, elementRegistry) {
 }
 
 HoverFix.$inject = [
+  'elementRegistry',
   'eventBus',
-  'dragging',
-  'elementRegistry'
+  'injector'
 ];
