+++ lib/features/hover-fix/index.js
@@ -1,0 +1,8 @@
+import HoverFix from './HoverFix';
+
+export default {
+  __init__: [
+    'hoverFix'
+  ],
+  hoverFix: [ 'type', HoverFix ],
+};
